Fix panics due to malformed input
Malformed BER input with lengths in the header
which are

(a) negative
(b) extremely large

would cause panics.

Changes:
* don't allocate slices of an untrusted length (memory exhaustion)
* avoid allocation of negative length slices
* return an int64 length from readPacket to make CopyN use clear
* add fuzzing corpus with cases which would previsouly panic

+++ header.go
@@ -5,19 +5,19 @@ import (
 	"io"
 )
 
-func readHeader(reader io.Reader) (identifier Identifier, length int, read int, err error) {
+func readHeader(reader io.Reader) (identifier Identifier, length int64, read int64, err error) {
 	if i, c, err := readIdentifier(reader); err != nil {
 		return Identifier{}, 0, read, err
 	} else {
 		identifier = i
-		read += c
+		read += int64(c)
 	}
 
 	if l, c, err := readLength(reader); err != nil {
 		return Identifier{}, 0, read, err
 	} else {
 		length = l
-		read += c
+		read += int64(c)
 	}
 
 	// Validate length type with identifier (x.600, 8.1.3.2.a)
